haitiironmarketapp/quickstart/admin/products/serializers.py
from dataclasses import fields
from itertools import product
from rest_framework import serializers

from products.models import Product

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = '__all__'